partition block manager & lsmtree gc fix, before test

+++ lower/linux_aio/linux_aio.c
@@ -95,7 +95,7 @@ void *poller(void *input) {
 				r=&done_array[i];
 				req=(algo_req*)r->data;
 				cb=r->obj;
-				if(r->res==-22){
+				if(r->res==(uint32_t)-22){
 					printf("error! %s %lu %llu\n",strerror(-r->res),r->res2,cb->u.c.offset);
 				}else if(r->res!=PAGESIZE){
 					printf("data size error %d!\n",errno);
